+++ src/cf-core/src/cf-vars-breakpoints.js
@@ -1,0 +1,29 @@
+/* ==========================================================================
+   Capital Framework
+   Breakpoint JavaScript variables.
+   Matched to values for use in JavaScript matched
+   to CSS breakpoints in cf-vars-breakpoints.less
+   All values are pixel based.
+   ========================================================================== */
+
+module.exports = {
+  bpXS: {
+    min: 0,
+    max: 600
+  },
+  bpSM: {
+    min: 601,
+    max: 900
+  },
+  bpMED: {
+    min: 901,
+    max: 1020
+  },
+  bpLG: {
+    min: 1021,
+    max: 1200
+  },
+  bpXL: {
+    min: 1201
+  }
+};